+++ packages/terra-i18n-plugin/tests/i18n-plugin-spec.test.js
@@ -79,6 +79,11 @@ describe('i18n-aggregator-plugin', () => {
       compiler._plugins['after-environment'][0]();
     });
 
+    beforeEach(() => {
+      // eslint-disable-next-line no-console
+      console.warn.mockClear();
+    });
+
     it('should create tranlsation files for all supported locales', () => {
       expect(supportedLocales.length).toEqual(outputtedFileName.length);
       expect(SpyOnWriteFileSync).toHaveBeenCalledTimes(supportedLocales.length);
@@ -141,6 +146,11 @@ describe('i18n-aggregator-plugin', () => {
         translationsDirectoryRouters: ['customFolder'],
       }).apply(compiler);
       compiler._plugins['after-environment'][0]();
+    });
+
+    beforeEach(() => {
+      // eslint-disable-next-line no-console
+      console.warn.mockClear();
     });
 
     it('should create tranlsation files for all supported locales', () => {
@@ -207,6 +217,11 @@ describe('i18n-aggregator-plugin', () => {
       compiler._plugins['after-environment'][0]();
     });
 
+    beforeEach(() => {
+      // eslint-disable-next-line no-console
+      console.warn.mockClear();
+    });
+
     it('should create tranlsation files for all supported locales', () => {
       expect(supportedLocales.length).toEqual(outputtedFileName.length);
       expect(SpyOnWriteFileSync).toHaveBeenCalledTimes(supportedLocales.length);
@@ -263,6 +278,11 @@ describe('i18n-aggregator-plugin', () => {
       compiler._plugins['after-environment'][0]();
     });
 
+    beforeEach(() => {
+      // eslint-disable-next-line no-console
+      console.warn.mockClear();
+    });
+
     it('should create tranlsation files for all supported locales', () => {
       expect(SpyOnInputReadDirSync).toHaveBeenCalledTimes(2);
       expect(SpyOnInputReadFileSync).toHaveBeenCalledTimes(supportedLocales.length);
@@ -308,6 +328,11 @@ describe('i18n-aggregator-plugin', () => {
         outputFileSystem,
       }).apply(compiler);
       compiler._plugins['after-environment'][0]();
+    });
+
+    beforeEach(() => {
+      // eslint-disable-next-line no-console
+      console.warn.mockClear();
     });
 
     it('should create tranlsation files for all supported locales', () => {